Version 0.63

+++ code/src/main/java/io/almostrealism/expression/Sine.java
@@ -21,6 +21,7 @@ import io.almostrealism.lang.LanguageOperations;
 
 import java.util.List;
 import java.util.OptionalInt;
+import java.util.OptionalLong;
 
 public class Sine extends Expression<Double> {
 	public Sine(Expression<Double> input) {
@@ -34,8 +35,8 @@ public class Sine extends Expression<Double> {
 	}
 
 	@Override
-	public OptionalInt upperBound(KernelStructureContext context) {
-		return OptionalInt.of(1);
+	public OptionalLong upperBound(KernelStructureContext context) {
+		return OptionalLong.of(1);
 	}
 
 	@Override
